requirements.txt
# model-finetune-tool 依赖文件
# 使用 uv 管理: uv sync
# 或 pip: pip install -r requirements.txt

# 核心依赖 (锁定版本范围以保证兼容性)
pydantic>=2.0,<3.0
pyyaml>=6.0,<7.0
openai>=1.0,<2.0
python-docx>=1.0,<2.0
pymupdf>=1.23,<2.0
markdown>=3.5,<4.0
sqlalchemy>=2.0,<3.0
pymysql>=1.1,<2.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
tqdm>=4.66,<5.0
click>=8.0,<9.0

# 开发依赖
pytest>=8.0,<9.0

# 可选依赖 (用于训练)
# transformers>=4.0,<5.0
# datasets>=2.0,<3.0
# peft>=0.4.0,<1.0
# accelerate>=0.20.0,<1.0
# bitsandbytes>=0.41.0,<1.0
